+++ src/main/scala/rps/models/GameResult.scala
@@ -1,0 +1,8 @@
+package rps.models
+
+sealed trait GameResult
+object GameResult {
+  case object Draw extends GameResult
+  case object UserWins extends GameResult
+  case object CpuWins extends GameResult
+}
